+++ run_integrated_tests.sh
@@ -4,8 +4,8 @@ echo "Linting..."
 poetry run task lint
 
 echo "Making test docker image..."
-#docker build -t local/cogment-py-sdk-integration-test --build-arg COGMENT_ORCHESTRATOR_IMAGE="local/orchestrator" -f integration_test.dockerfile .
-docker build -t local/cogment-py-sdk-integration-test --build-arg COGMENT_ORCHESTRATOR_IMAGE="cogment/orchestrator:latest" -f integration_test.dockerfile .
+docker build -t local/cogment-py-sdk-integration-test --build-arg COGMENT_ORCHESTRATOR_IMAGE="local/orchestrator" -f integration_test.dockerfile .
+#docker build -t local/cogment-py-sdk-integration-test --build-arg COGMENT_ORCHESTRATOR_IMAGE="cogment/orchestrator:latest" -f integration_test.dockerfile .
 #docker build -t local/cogment-py-sdk-integration-test --build-arg COGMENT_ORCHESTRATOR_IMAGE="registry.gitlab.com/ai-r/cogment-orchestrator:latest" -f integration_test.dockerfile .
 
 echo "Running test..."