+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+
+- New type of controller
+
+### Added
+
+- WatchTrials function
+
 ## v1.0.0-alpha8 - 2021-02-26
 
 ### Changed